requirements.txt
Flask~=2.0.2
msrest~=0.6.21
Werkzeug~=2.0.2

azure-cognitiveservices-vision-computervision~=0.9.0